refactor: reorganize repository layout
