+++ README.md
@@ -61,9 +61,9 @@ This avoids the old runtime hazards:
 - no full all-pairs cosine similarity matrix is allocated,
 - raw IMDb TSVs are never read by the app runtime.
 
-The legacy `MovieRecommender` class remains in `src/movie_recommender.py` for
-compatibility and small tests, but it should not be used for full or preview
-runtime serving.
+The legacy `MovieRecommender` class remains in
+`src/movie_recommender/recommenders/legacy_content.py` for compatibility and
+small tests, but it should not be used for full or preview runtime serving.
 
 Build or refresh the SQLite store by running a recommendation once:
