+++ src/movie_recommender/__init__.py
@@ -1,19 +1,19 @@
+"""Reusable MovieRecommender package."""
+
 __all__ = ["MovieDatasetReducer", "MovieRecommender", "SQLiteMovieRecommender"]
 
 
 def __getattr__(name):
     if name == "MovieDatasetReducer":
-        from .movie_recommender.data.dataset_reducer import MovieDatasetReducer
+        from .data.dataset_reducer import MovieDatasetReducer
 
         return MovieDatasetReducer
     if name == "MovieRecommender":
-        from .movie_recommender.recommenders.legacy_content import MovieRecommender
+        from .recommenders.legacy_content import MovieRecommender
 
         return MovieRecommender
     if name == "SQLiteMovieRecommender":
-        from .movie_recommender.recommenders.sqlite_recommender import (
-            SQLiteMovieRecommender,
-        )
+        from .recommenders.sqlite_recommender import SQLiteMovieRecommender
 
         return SQLiteMovieRecommender
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
