+++ src/movie_recommender/cli/imdb_bootstrap.py
@@ -1,0 +1,5 @@
+"""Command line entrypoint for IMDb source bootstrap helpers."""
+
+from movie_recommender.data.imdb_bootstrap import main
+
+__all__ = ["main"]
