+++ src/movie_recommender/data/__init__.py
@@ -1,0 +1,5 @@
+"""Dataset loading and reduction helpers."""
+
+from .dataset_reducer import MovieDatasetReducer
+
+__all__ = ["MovieDatasetReducer"]
